+++ tests/e2e/helpers/weather-functions.js
@@ -1,7 +1,5 @@
 const helpers = require("./global-setup");
-const path = require("path");
-const fs = require("fs");
-const { generateWeather, generateWeatherForecast, generateWeatherHourly } = require("../../mocks/weather_test");
+const { injectMockData } = require("../../utils/weather_mocker");
 
 exports.getText = async (element, result) => {
 	const elem = await helpers.waitForElement(element);
@@ -14,18 +12,8 @@ exports.getText = async (element, result) => {
 	).toBe(result);
 };
 
-exports.startApp = async (configFile, additionalMockData) => {
-	let mockWeather;
-	if (configFile.includes("forecast")) {
-		mockWeather = generateWeatherForecast(additionalMockData);
-	} else if (configFile.includes("hourly")) {
-		mockWeather = generateWeatherHourly(additionalMockData);
-	} else {
-		mockWeather = generateWeather(additionalMockData);
-	}
-	let content = fs.readFileSync(path.resolve(__dirname + "../../../../" + configFile)).toString();
-	content = content.replace("#####WEATHERDATA#####", mockWeather);
-	fs.writeFileSync(path.resolve(__dirname + "../../../../config/config.js"), content);
+exports.startApp = async (configFileName, additionalMockData) => {
+	injectMockData(configFileName, additionalMockData);
 	await helpers.startApplication("");
 	await helpers.getDocument();
 };